pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentmx"
version = "0.1.0"
description = "AgentM-X – Autonomous, Self-Evolving Multi-Agent System (Windows, Full Host Control)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "AgentM-X"}]
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "typer>=0.12",
  "httpx>=0.27",
  "loguru>=0.7",
  "psutil>=5.9",
  "playwright>=1.45",
  "pillow>=10.3",
  "pytesseract>=0.3.10",
  "pyautogui>=0.9.54",
  "pywinauto>=0.6.8",
]

[project.scripts]
agentmx = "agentmx.cli:main"

[tool.uv]
dev-dependencies = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
]
